Cale/latest edx lint

+++ pavelib/quality.py
@@ -57,27 +57,64 @@ def find_fixme(options):
 
         apps_list = ' '.join(top_python_dirs(system))
 
-        pythonpath_prefix = (
-            "PYTHONPATH={system}/djangoapps:common/djangoapps:common/lib".format(
-                system=system
-            )
-        )
-
-        sh(
-            "{pythonpath_prefix} pylint --disable R,C,W,E --enable=fixme "
-            "--msg-template={msg_template} {apps} "
-            "| tee {report_dir}/pylint_fixme.report".format(
-                pythonpath_prefix=pythonpath_prefix,
-                msg_template='"{path}:{line}: [{msg_id}({symbol}), {obj}] {msg}"',
+        cmd = (
+            "pylint --disable all --enable=fixme "
+            "--output-format=parseable {apps} "
+            "> {report_dir}/pylint_fixme.report".format(
                 apps=apps_list,
                 report_dir=report_dir
             )
         )
 
+        sh(cmd, ignore_error=True)
+
         num_fixme += _count_pylint_violations(
             "{report_dir}/pylint_fixme.report".format(report_dir=report_dir))
 
     print "Number of pylint fixmes: " + str(num_fixme)
+
+
+def _get_pylint_violations(systems=ALL_SYSTEMS.split(','), errors_only=False, clean=True):
+    """
+    Runs pylint. Returns a tuple of (number_of_violations, list_of_violations)
+    where list_of_violations is a list of all pylint violations found, separated
+    by new lines.
+    """
+    # Make sure the metrics subdirectory exists
+    Env.METRICS_DIR.makedirs_p()
+
+    num_violations = 0
+    violations_list = []
+
+    for system in systems:
+        # Directory to put the pylint report in.
+        # This makes the folder if it doesn't already exist.
+        report_dir = (Env.REPORT_DIR / system).makedirs_p()
+
+        flags = []
+        if errors_only:
+            flags.append("--errors-only")
+
+        apps_list = ' '.join(top_python_dirs(system))
+
+        system_report = report_dir / 'pylint.report'
+        if clean or not system_report.exists():
+            sh(
+                "pylint {flags} --output-format=parseable {apps} "
+                "> {report_dir}/pylint.report".format(
+                    flags=" ".join(flags),
+                    apps=apps_list,
+                    report_dir=report_dir
+                ),
+                ignore_error=True,
+            )
+
+        num_violations += _count_pylint_violations(system_report)
+        with open(system_report) as report_contents:
+            violations_list.extend(report_contents)
+
+    # Print number of violations to log
+    return num_violations, violations_list
 
 
 @task
@@ -94,41 +131,10 @@ def run_pylint(options):
     fail the task if too many violations are found.
     """
     lower_violations_limit, upper_violations_limit, errors, systems = _parse_pylint_options(options)
+    errors = getattr(options, 'errors', False)
+    systems = getattr(options, 'system', ALL_SYSTEMS).split(',')
 
-    # Make sure the metrics subdirectory exists
-    Env.METRICS_DIR.makedirs_p()
-
-    num_violations = 0
-    for system in systems:
-        # Directory to put the pylint report in.
-        # This makes the folder if it doesn't already exist.
-        report_dir = (Env.REPORT_DIR / system).makedirs_p()
-
-        flags = []
-        if errors:
-            flags.append("--errors-only")
-
-        apps_list = ' '.join(top_python_dirs(system))
-
-        pythonpath_prefix = (
-            "PYTHONPATH={system}/djangoapps:common/djangoapps:common/lib".format(
-                system=system
-            )
-        )
-
-        sh(
-            "{pythonpath_prefix} pylint {flags} --msg-template={msg_template} {apps} | "
-            "tee {report_dir}/pylint.report".format(
-                pythonpath_prefix=pythonpath_prefix,
-                flags=" ".join(flags),
-                msg_template='"{path}:{line}: [{msg_id}({symbol}), {obj}] {msg}"',
-                apps=apps_list,
-                report_dir=report_dir
-            )
-        )
-
-        num_violations += _count_pylint_violations(
-            "{report_dir}/pylint.report".format(report_dir=report_dir))
+    num_violations, violations_list = _get_pylint_violations(systems, errors)
 
     # Print number of violations to log
     violations_count_str = "Number of pylint violations: " + str(num_violations)
@@ -198,37 +204,35 @@ def _count_pylint_violations(report_file):
     return num_violations_report
 
 
-def _get_pep8_violations():
+def _get_pep8_violations(clean=True):
     """
     Runs pep8. Returns a tuple of (number_of_violations, violations_string)
     where violations_string is a string of all pep8 violations found, separated
     by new lines.
     """
     report_dir = (Env.REPORT_DIR / 'pep8')
-    report_dir.rmtree(ignore_errors=True)
+    if clean:
+        report_dir.rmtree(ignore_errors=True)
     report_dir.makedirs_p()
+    report = report_dir / 'pep8.report'
 
     # Make sure the metrics subdirectory exists
     Env.METRICS_DIR.makedirs_p()
 
-    sh('pep8 . | tee {report_dir}/pep8.report -a'.format(report_dir=report_dir))
+    if not report.exists():
+        sh('pep8 . | tee {} -a'.format(report))
 
-    count, violations_list = _pep8_violations(
-        "{report_dir}/pep8.report".format(report_dir=report_dir)
-    )
+    violations_list = _pep8_violations(report)
 
-    return count, violations_list
+    return (len(violations_list), violations_list)
 
 
 def _pep8_violations(report_file):
     """
-    Returns a tuple of (num_violations, violations_list) for all
-    pep8 violations in the given report_file.
+    Returns the list of all pep8 violations in the given report_file.
     """
     with open(report_file) as f:
-        violations_list = f.readlines()
-    num_lines = len(violations_list)
-    return num_lines, violations_list
+        return f.readlines()
 
 
 @task
@@ -680,6 +684,7 @@ def _get_xsscommitlint_count(filename):
 @cmdopts([
     ("compare-branch=", "b", "Branch to compare against, defaults to origin/master"),
     ("percentage=", "p", "fail if diff-quality is below this percentage"),
+    ("limit=", "l", "Limits for number of acceptable violations - either <upper> or <lower>:<upper>"),
 ])
 @timed
 def run_quality(options):
@@ -698,7 +703,7 @@ def run_quality(options):
     # Save the pass variable. It will be set to false later if failures are detected.
     diff_quality_percentage_pass = True
 
-    def _pep8_output(count, violations_list, is_html=False):
+    def _lint_output(linter, count, violations_list, is_html=False, limit=0):
         """
         Given a count & list of pep8 violations, pretty-print the pep8 output.
         If `is_html`, will print out with HTML markup.
@@ -706,26 +711,26 @@ def run_quality(options):
         if is_html:
             lines = ['<body>\n']
             sep = '-------------<br/>\n'
-            title = "<h1>Quality Report: pep8</h1>\n"
+            title = HTML("<h1>Quality Report: {}</h1>\n").format(linter)
             violations_bullets = ''.join(
                 [HTML('<li>{violation}</li><br/>\n').format(violation=violation) for violation in violations_list]
             )
             violations_str = HTML('<ul>\n{bullets}</ul>\n').format(bullets=HTML(violations_bullets))
-            violations_count_str = "<b>Violations</b>: {count}<br/>\n"
-            fail_line = "<b>FAILURE</b>: pep8 count should be 0<br/>\n"
+            violations_count_str = HTML("<b>Violations</b>: {count}<br/>\n")
+            fail_line = HTML("<b>FAILURE</b>: {} count should be 0<br/>\n").format(linter)
         else:
             lines = []
             sep = '-------------\n'
-            title = "Quality Report: pep8\n"
+            title = "Quality Report: {}\n".format(linter)
             violations_str = ''.join(violations_list)
             violations_count_str = "Violations: {count}\n"
-            fail_line = "FAILURE: pep8 count should be 0\n"
+            fail_line = "FAILURE: {} count should be {}\n".format(linter, limit)
 
         violations_count_str = violations_count_str.format(count=count)
 
         lines.extend([sep, title, sep, violations_str, sep, violations_count_str])
 
-        if count > 0:
+        if count > limit > -1:
             lines.append(fail_line)
         lines.append(sep + '\n')
         if is_html:
@@ -734,16 +739,34 @@ def run_quality(options):
         return ''.join(lines)
 
     # Run pep8 directly since we have 0 violations on master
-    (count, violations_list) = _get_pep8_violations()
+    (count, violations_list) = _get_pep8_violations(clean=False)
 
     # Print number of violations to log
-    print _pep8_output(count, violations_list)
+    print _lint_output('pep8', count, violations_list)
 
     # Also write the number of violations to a file
     with open(dquality_dir / "diff_quality_pep8.html", "w") as f:
-        f.write(_pep8_output(count, violations_list, is_html=True))
+        f.write(_lint_output('pep8', count, violations_list, is_html=True))
 
     if count > 0:
+        diff_quality_percentage_pass = False
+
+    # Generate diff-quality html report for pylint, and print to console
+    # If pylint reports exist, use those
+    # Otherwise, `diff-quality` will call pylint itself
+
+    (count, violations_list) = _get_pylint_violations(clean=False)
+
+    _, upper_violations_limit, _, _ = _parse_pylint_options(options)
+
+    # Print number of violations to log
+    print _lint_output('pylint', count, violations_list, limit=upper_violations_limit)
+
+    # Also write the number of violations to a file
+    with open(dquality_dir / "diff_quality_pylint.html", "w") as f:
+        f.write(_lint_output('pylint', count, violations_list, is_html=True, limit=upper_violations_limit))
+
+    if count > upper_violations_limit > -1:
         diff_quality_percentage_pass = False
 
     # ----- Set up for diff-quality pylint call -----
@@ -759,36 +782,12 @@ def run_quality(options):
     if diff_threshold > -1:
         percentage_string = u'--fail-under={0}'.format(diff_threshold)
 
-    # Generate diff-quality html report for pylint, and print to console
-    # If pylint reports exist, use those
-    # Otherwise, `diff-quality` will call pylint itself
-
-    pylint_files = get_violations_reports("pylint")
-    pylint_reports = u' '.join(pylint_files)
-
     eslint_files = get_violations_reports("eslint")
     eslint_reports = u' '.join(eslint_files)
-
-    pythonpath_prefix = (
-        "PYTHONPATH=$PYTHONPATH:lms:lms/djangoapps:cms:cms/djangoapps:"
-        "common:common/djangoapps:common/lib"
-    )
-
-    # run diff-quality for pylint.
-    if not run_diff_quality(
-            violations_type="pylint",
-            prefix=pythonpath_prefix,
-            reports=pylint_reports,
-            percentage_string=percentage_string,
-            branch_string=compare_branch_string,
-            dquality_dir=dquality_dir
-    ):
-        diff_quality_percentage_pass = False
 
     # run diff-quality for eslint.
     if not run_diff_quality(
             violations_type="eslint",
-            prefix=pythonpath_prefix,
             reports=eslint_reports,
             percentage_string=percentage_string,
             branch_string=compare_branch_string,
@@ -803,7 +802,7 @@ def run_quality(options):
 
 
 def run_diff_quality(
-        violations_type=None, prefix=None, reports=None, percentage_string=None, branch_string=None, dquality_dir=None
+        violations_type=None, reports=None, percentage_string=None, branch_string=None, dquality_dir=None
 ):
     """
     This executes the diff-quality commandline tool for the given violation type (e.g., pylint, eslint).
@@ -812,11 +811,10 @@ def run_diff_quality(
     """
     try:
         sh(
-            "{pythonpath_prefix} diff-quality --violations={type} "
+            "diff-quality --violations={type} "
             "{reports} {percentage_string} {compare_branch_string} "
             "--html-report {dquality_dir}/diff_quality_{type}.html ".format(
                 type=violations_type,
-                pythonpath_prefix=prefix,
                 reports=reports,
                 percentage_string=percentage_string,
                 compare_branch_string=branch_string,
